+++ sdk/c4-client.ts
@@ -2,8 +2,8 @@
 // Generated from /openapi.json via src/openapi-sdk-gen.js.
 // Do not edit by hand — re-run `c4 openapi --sdk` to refresh.
 
-// Spec version: 1.10.111
-// Generated at: 2026-05-02T16:20:34.213Z
+// Spec version: 1.10.112
+// Generated at: 2026-05-02T16:31:38.077Z
 
 export interface postAuthLoginBody {
   user: string; /** Username */